clean up file
