+++ cellpack/autopack/utils.py
@@ -3,3 +3,114 @@ import numpy
 
 def get_distance(pt1, pt2):
     return numpy.linalg.norm(pt2 - pt1)
+
+
+def ingredient_compare1(x, y):
+    """
+    sort ingredients using decreasing priority and decreasing radii for
+    priority ties and decreasing completion for radii ties
+    for priority > 0
+    """
+    p1 = x.packingPriority
+    p2 = y.packingPriority
+    if p1 < p2:  # p1 > p2
+        return 1
+    elif p1 == p2:  # p1 == p1
+        r1 = x.minRadius
+        r2 = y.minRadius
+        if r1 > r2:  # r1 < r2
+            return 1
+        elif r1 == r2:  # r1 == r2
+            c1 = x.completion
+            c2 = y.completion
+            if c1 > c2:  # c1 > c2
+                return 1
+            elif c1 == c2:
+                return 0
+            else:
+                return -1
+        else:
+            return -1
+    else:
+        return -1
+
+
+def ingredient_compare0(x, y):
+    """
+    sort ingredients using decreasing priority and decreasing radii for
+    priority ties and decreasing completion for radii ties
+    for priority < 0
+    """
+    p1 = x.packingPriority
+    p2 = y.packingPriority
+    if p1 > p2:  # p1 > p2
+        return 1
+    elif p1 == p2:  # p1 == p1
+        r1 = x.minRadius
+        r2 = y.minRadius
+        if r1 > r2:  # r1 < r2
+            return 1
+        elif r1 == r2:  # r1 == r2
+            c1 = x.completion
+            c2 = y.completion
+            if c1 > c2:  # c1 > c2
+                return 1
+            elif c1 == c2:
+                return 0
+            else:
+                return -1
+        else:
+            return -1
+    else:
+        return -1
+
+
+def ingredient_compare2(x, y):
+    """
+    sort ingredients using decreasing radii and decresing completion
+    for radii matches:
+    priority = 0
+    """
+    c1 = x.minRadius
+    c2 = y.minRadius
+    if c1 < c2:
+        return 1
+    elif c1 == c2:
+        r1 = x.completion
+        r2 = y.completion
+        if r1 > r2:
+            return 1
+        elif r1 == r2:
+            return 0
+        else:
+            return -1
+    else:  # x < y
+        return -1
+
+
+def cmp_to_key(mycmp):
+    "Convert a cmp= function into a key= function"
+
+    class K:
+        def __init__(self, obj, *args):
+            self.obj = obj
+
+        def __lt__(self, other):
+            return mycmp(self.obj, other.obj) < 0
+
+        def __gt__(self, other):
+            return mycmp(self.obj, other.obj) > 0
+
+        def __eq__(self, other):
+            return mycmp(self.obj, other.obj) == 0
+
+        def __le__(self, other):
+            return mycmp(self.obj, other.obj) <= 0
+
+        def __ge__(self, other):
+            return mycmp(self.obj, other.obj) >= 0
+
+        def __ne__(self, other):
+            return mycmp(self.obj, other.obj) != 0
+
+    return K
